+++ src.ts/ethers.ts
@@ -49,6 +49,7 @@ export {
 export {
     id,
     ensNormalize, isValidName, namehash, dnsEncode,
+    hashAuthorization, verifyAuthorization,
     hashMessage, verifyMessage,
     solidityPacked, solidityPackedKeccak256, solidityPackedSha256,
     TypedDataEncoder,
@@ -69,9 +70,9 @@ export {
 
     BrowserProvider,
 
-    AlchemyProvider, AnkrProvider, ChainstackProvider, CloudflareProvider,
-    EtherscanProvider, InfuraProvider, InfuraWebSocketProvider, PocketProvider,
-    QuickNodeProvider,
+    AlchemyProvider, AnkrProvider, BlockscoutProvider, ChainstackProvider,
+    CloudflareProvider, EtherscanProvider, InfuraProvider,
+    InfuraWebSocketProvider, PocketProvider, QuickNodeProvider,
 
     IpcSocketProvider, SocketProvider, WebSocketProvider,
 
@@ -89,7 +90,7 @@ export {
 } from "./providers/index.js";
 
 export {
-    accessListify,
+    accessListify, authorizationify,
     computeAddress, recoverAddress,
     Transaction
 } from "./transaction/index.js";
@@ -159,26 +160,33 @@ export type {
 
 export type { ProgressCallback, SignatureLike } from "./crypto/index.js";
 
-export type { TypedDataDomain, TypedDataField } from "./hash/index.js";
+export type {
+    AuthorizationRequest, TypedDataDomain, TypedDataField
+} from "./hash/index.js";
 
 export type {
     Provider, Signer,
 
+    CommunityResourcable,
+
     AbstractProviderOptions, BrowserProviderOptions, FallbackProviderOptions,
 
-    AbstractProviderPlugin, BlockParams, BlockTag, ContractRunner, DebugEventBrowserProvider,
-    Eip1193Provider, EventFilter, Filter, FilterByBlockHash, GasCostParameters,
-    JsonRpcApiProviderOptions, JsonRpcError, JsonRpcPayload, JsonRpcResult,
-    JsonRpcTransactionRequest, LogParams, MinedBlock, MinedTransactionResponse, Networkish,
-    OrphanFilter, PerformActionFilter, PerformActionRequest, PerformActionTransaction,
-    PreparedTransactionRequest, ProviderEvent, Subscriber, Subscription, TopicFilter,
-    TransactionReceiptParams, TransactionRequest, TransactionResponseParams,
-    WebSocketCreator, WebSocketLike
+    AbstractProviderPlugin, BlockParams, BlockTag, BrowserDiscoverOptions,
+    ContractRunner, DebugEventBrowserProvider, Eip1193Provider,
+    Eip6963ProviderInfo, EventFilter, Filter, FilterByBlockHash,
+    GasCostParameters, JsonRpcApiProviderOptions, JsonRpcError,
+    JsonRpcPayload, JsonRpcResult, JsonRpcTransactionRequest, LogParams,
+    MinedBlock, MinedTransactionResponse, Networkish, OrphanFilter,
+    PerformActionFilter, PerformActionRequest, PerformActionTransaction,
+    PreparedTransactionRequest, ProviderEvent, Subscriber, Subscription,
+    TopicFilter, TransactionReceiptParams, TransactionRequest,
+    TransactionResponseParams, WebSocketCreator, WebSocketLike
 } from "./providers/index.js";
 
 export type {
     AccessList, AccessListish, AccessListEntry,
-    Blob, BlobLike, KzgLibrary,
+    Authorization, AuthorizationLike,
+    Blob, BlobLike, KzgLibrary, KzgLibraryLike,
     TransactionLike
 } from "./transaction/index.js";