+++ src.ts/transaction/transaction.ts
@@ -10,12 +10,15 @@ import {
 } from "../utils/index.js";
 
 import { accessListify } from "./accesslist.js";
+import { authorizationify } from "./authorization.js";
 import { recoverAddress } from "./address.js";
 
 import type { BigNumberish, BytesLike } from "../utils/index.js";
 import type { SignatureLike } from "../crypto/index.js";
 
-import type { AccessList, AccessListish } from "./index.js";
+import type {
+    AccessList, AccessListish, Authorization, AuthorizationLike
+} from "./index.js";
 
 
 const BN_0 = BigInt(0);
@@ -128,7 +131,12 @@ export interface TransactionLike<A = string> {
      *  This is generally ``null``, unless you are creating BLOb
      *  transactions.
      */
-    kzg?: null | KzgLibrary;
+    kzg?: null | KzgLibraryLike;
+
+    /**
+     *  The [[link-eip-7702]] authorizations (if any).
+     */
+    authorizationList?: null | Array<Authorization>;
 }
 
 /**
@@ -165,6 +173,74 @@ export interface KzgLibrary {
     computeBlobKzgProof: (blob: Uint8Array, commitment: Uint8Array) => Uint8Array;
 }
 
+/**
+ *  A KZG Library with any of the various API configurations.
+ *  As the library is still experimental and the API is not
+ *  stable, depending on the version used the method names and
+ *  signatures are still in flux.
+ *
+ *  This allows any of the versions to be passed into Transaction
+ *  while providing a stable external API.
+ */
+export type KzgLibraryLike  = KzgLibrary | {
+    // kzg-wasm >= 0.5.0
+    blobToKZGCommitment: (blob: string) => string;
+    computeBlobKZGProof: (blob: string, commitment: string) => string;
+} | {
+    // micro-ecc-signer
+    blobToKzgCommitment: (blob: string) => string | Uint8Array;
+    computeBlobProof: (blob: string, commitment: string) => string | Uint8Array;
+};
+
+function getKzgLibrary(kzg: KzgLibraryLike): KzgLibrary {
+
+    const blobToKzgCommitment = (blob: Uint8Array) => {
+
+        if ("computeBlobProof" in kzg) {
+            // micro-ecc-signer; check for computeBlobProof since this API
+            // expects a string while the kzg-wasm below expects a Unit8Array
+
+            if ("blobToKzgCommitment" in kzg && typeof(kzg.blobToKzgCommitment) === "function") {
+                return getBytes(kzg.blobToKzgCommitment(hexlify(blob)))
+            }
+
+        } else if ("blobToKzgCommitment" in kzg && typeof(kzg.blobToKzgCommitment) === "function") {
+            // kzg-wasm <0.5.0; blobToKzgCommitment(Uint8Array) => Uint8Array
+
+            return getBytes(kzg.blobToKzgCommitment(blob));
+        }
+
+        // kzg-wasm >= 0.5.0; blobToKZGCommitment(string) => string
+        if ("blobToKZGCommitment" in kzg && typeof(kzg.blobToKZGCommitment) === "function") {
+            return getBytes(kzg.blobToKZGCommitment(hexlify(blob)));
+        }
+
+        assertArgument(false, "unsupported KZG library", "kzg", kzg);
+    };
+
+    const computeBlobKzgProof = (blob: Uint8Array, commitment: Uint8Array) => {
+
+        // micro-ecc-signer
+        if ("computeBlobProof" in kzg && typeof(kzg.computeBlobProof) === "function") {
+            return getBytes(kzg.computeBlobProof(hexlify(blob), hexlify(commitment)))
+        }
+
+        // kzg-wasm <0.5.0; computeBlobKzgProof(Uint8Array, Uint8Array) => Uint8Array
+        if ("computeBlobKzgProof" in kzg && typeof(kzg.computeBlobKzgProof) === "function") {
+            return kzg.computeBlobKzgProof(blob, commitment);
+        }
+
+        // kzg-wasm >= 0.5.0; computeBlobKZGProof(string, string) => string
+        if ("computeBlobKZGProof" in kzg && typeof(kzg.computeBlobKZGProof) === "function") {
+            return getBytes(kzg.computeBlobKZGProof(hexlify(blob), hexlify(commitment)));
+        }
+
+        assertArgument(false, "unsupported KZG library", "kzg", kzg);
+    };
+
+    return { blobToKzgCommitment, computeBlobKzgProof };
+}
+
 function getVersionedHash(version: number, hash: BytesLike): string {
     let versioned = version.toString(16);
     while (versioned.length < 2) { versioned = "0" + versioned; }
@@ -180,6 +256,32 @@ function handleAddress(value: string): null | string {
 function handleAccessList(value: any, param: string): AccessList {
     try {
         return accessListify(value);
+    } catch (error: any) {
+        assertArgument(false, error.message, param, value);
+    }
+}
+
+function handleAuthorizationList(value: any, param: string): Array<Authorization> {
+    try {
+        if (!Array.isArray(value)) { throw new Error("authorizationList: invalid array"); }
+        const result: Array<Authorization> = [ ];
+        for (let i = 0; i < value.length; i++) {
+            const auth: Array<string> = value[i];
+            if (!Array.isArray(auth)) { throw new Error(`authorization[${ i }]: invalid array`); }
+            if (auth.length !== 6) { throw new Error(`authorization[${ i }]: wrong length`); }
+            if (!auth[1]) { throw new Error(`authorization[${ i }]: null address`); }
+            result.push({
+                address: <string>handleAddress(auth[1]),
+                nonce: handleUint(auth[2], "nonce"),
+                chainId: handleUint(auth[0], "chainId"),
+                signature: Signature.from({
+                    yParity: <0 | 1>handleNumber(auth[3], "yParity"),
+                    r: zeroPadValue(auth[4], 32),
+                    s: zeroPadValue(auth[5], 32)
+                })
+            });
+        }
+        return result;
     } catch (error: any) {
         assertArgument(false, error.message, param, value);
     }
@@ -206,6 +308,19 @@ function formatNumber(_value: BigNumberish, name: string): Uint8Array {
 
 function formatAccessList(value: AccessListish): Array<[ string, Array<string> ]> {
     return accessListify(value).map((set) => [ set.address, set.storageKeys ]);
+}
+
+function formatAuthorizationList(value: Array<Authorization>): Array<Array<string | Uint8Array>> {
+    return value.map((a) => {
+        return [
+            formatNumber(a.chainId, "chainId"),
+            a.address,
+            formatNumber(a.nonce, "nonce"),
+            formatNumber(a.signature.yParity, "yParity"),
+            a.signature.r,
+            a.signature.s
+        ];
+    });
 }
 
 function formatHashes(value: Array<string>, param: string): Array<string> {
@@ -548,6 +663,58 @@ function _serializeEip4844(tx: Transaction, sig: null | Signature, blobs: null |
     return concat([ "0x03", encodeRlp(fields)]);
 }
 
+function _parseEip7702(data: Uint8Array): TransactionLike {
+    const fields: any = decodeRlp(getBytes(data).slice(1));
+
+    assertArgument(Array.isArray(fields) && (fields.length === 10 || fields.length === 13),
+        "invalid field count for transaction type: 4", "data", hexlify(data));
+
+    const tx: TransactionLike = {
+        type:                  4,
+        chainId:               handleUint(fields[0], "chainId"),
+        nonce:                 handleNumber(fields[1], "nonce"),
+        maxPriorityFeePerGas:  handleUint(fields[2], "maxPriorityFeePerGas"),
+        maxFeePerGas:          handleUint(fields[3], "maxFeePerGas"),
+        gasPrice:              null,
+        gasLimit:              handleUint(fields[4], "gasLimit"),
+        to:                    handleAddress(fields[5]),
+        value:                 handleUint(fields[6], "value"),
+        data:                  hexlify(fields[7]),
+        accessList:            handleAccessList(fields[8], "accessList"),
+        authorizationList:     handleAuthorizationList(fields[9], "authorizationList"),
+    };
+
+    // Unsigned EIP-7702 Transaction
+    if (fields.length === 10) { return tx; }
+
+    _parseEipSignature(tx, fields.slice(10));
+
+    return tx;
+}
+
+function _serializeEip7702(tx: Transaction, sig: null | Signature): string {
+    const fields: Array<any> = [
+        formatNumber(tx.chainId, "chainId"),
+        formatNumber(tx.nonce, "nonce"),
+        formatNumber(tx.maxPriorityFeePerGas || 0, "maxPriorityFeePerGas"),
+        formatNumber(tx.maxFeePerGas || 0, "maxFeePerGas"),
+        formatNumber(tx.gasLimit, "gasLimit"),
+        (tx.to || "0x"),
+        formatNumber(tx.value, "value"),
+        tx.data,
+        formatAccessList(tx.accessList || [ ]),
+        formatAuthorizationList(tx.authorizationList || [ ])
+    ];
+
+    if (sig) {
+        fields.push(formatNumber(sig.yParity, "yParity"));
+        fields.push(toBeArray(sig.r));
+        fields.push(toBeArray(sig.s));
+    }
+
+    return concat([ "0x04", encodeRlp(fields)]);
+}
+
 /**
  *  A **Transaction** describes an operation to be executed on
  *  Ethereum by an Externally Owned Account (EOA). It includes
@@ -578,6 +745,7 @@ export class Transaction implements TransactionLike<string> {
     #blobVersionedHashes: null | Array<string>;
     #kzg: null | KzgLibrary;
     #blobs: null | Array<Blob>;
+    #auths: null | Array<Authorization>;
 
     /**
      *  The transaction type.
@@ -603,6 +771,9 @@ export class Transaction implements TransactionLike<string> {
             case 3: case "cancun": case "eip-4844":
                 this.#type = 3;
                 break;
+            case 4: case "pectra": case "eip-7702":
+                this.#type = 4;
+                break;
             default:
                 assertArgument(false, "unsupported transaction type", "type", value);
         }
@@ -617,6 +788,7 @@ export class Transaction implements TransactionLike<string> {
             case 1: return "eip-2930";
             case 2: return "eip-1559";
             case 3: return "eip-4844";
+            case 4: return "eip-7702";
         }
 
         return null;
@@ -745,6 +917,22 @@ export class Transaction implements TransactionLike<string> {
         this.#accessList = (value == null) ? null: accessListify(value);
     }
 
+    get authorizationList(): null | Array<Authorization> {
+        const value = this.#auths || null;
+        if (value == null) {
+            if (this.type === 4) {
+                // @TODO: in v7, this should become a live object itself,
+                // otherwise mutation is inconsistent
+                return [ ];
+            }
+        }
+        return value;
+    }
+    set authorizationList(auths: null | Array<AuthorizationLike>) {
+        this.#auths = (auths == null) ? null: auths.map((a) =>
+          authorizationify(a));
+    }
+
     /**
      *  The max fee per blob gas for Cancun transactions.
      */
@@ -862,8 +1050,12 @@ export class Transaction implements TransactionLike<string> {
     }
 
     get kzg(): null | KzgLibrary { return this.#kzg; }
-    set kzg(kzg: null | KzgLibrary) {
-        this.#kzg = kzg;
+    set kzg(kzg: null | KzgLibraryLike) {
+        if (kzg == null) {
+            this.#kzg = null;
+        } else {
+            this.#kzg = getKzgLibrary(kzg);
+        }
     }
 
     /**
@@ -884,8 +1076,9 @@ export class Transaction implements TransactionLike<string> {
         this.#accessList = null;
         this.#maxFeePerBlobGas = null;
         this.#blobVersionedHashes = null;
-        this.#blobs = null;
         this.#kzg = null;
+        this.#blobs = null;
+        this.#auths = null;
     }
 
     /**
@@ -945,6 +1138,8 @@ export class Transaction implements TransactionLike<string> {
                 return _serializeEip1559(this, sig);
             case 3:
                 return _serializeEip4844(this, sig, sidecar ? this.blobs: null);
+            case 4:
+                return _serializeEip7702(this, sig);
         }
 
         assert(false, "unsupported transaction type", "UNSUPPORTED_OPERATION", { operation: ".serialized" });
@@ -1018,7 +1213,9 @@ export class Transaction implements TransactionLike<string> {
             types.push(this.type);
 
         } else {
-            if (hasFee) {
+            if (this.authorizationList && this.authorizationList.length) {
+                types.push(4);
+            } else if (hasFee) {
                 types.push(2);
             } else if (hasGasPrice) {
                 types.push(1);
@@ -1136,6 +1333,7 @@ export class Transaction implements TransactionLike<string> {
                 case 1: return Transaction.from(_parseEip2930(payload));
                 case 2: return Transaction.from(_parseEip1559(payload));
                 case 3: return Transaction.from(_parseEip4844(payload));
+                case 4: return Transaction.from(_parseEip7702(payload));
             }
             assert(false, "unsupported transaction type", "UNSUPPORTED_OPERATION", { operation: "from" });
         }
@@ -1154,6 +1352,9 @@ export class Transaction implements TransactionLike<string> {
         if (tx.chainId != null) { result.chainId = tx.chainId; }
         if (tx.signature != null) { result.signature = Signature.from(tx.signature); }
         if (tx.accessList != null) { result.accessList = tx.accessList; }
+        if (tx.authorizationList != null) {
+            result.authorizationList = tx.authorizationList;
+        }
 
         // This will get overwritten by blobs, if present
         if (tx.blobVersionedHashes != null) { result.blobVersionedHashes = tx.blobVersionedHashes; }